requirements.txt
django==4.2
djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0
googlemaps==4.10.0
setuptools<81
